Add files via upload
Hey guys,

I've used Nabel's ResDate to help me create a nice LocalDate in Room menu. Still mainly having trouble with the displayBooking() method. 

I think theres an issue with the rooms to be booked array, and I think that if we can sort that we'll be there. I'll try to sort it in morning before work :)